+++ source/extensions/omni.isaac.gxf_bridge/nodes/OgnGXFPublishImage.cpp
@@ -29,6 +29,8 @@ public:
             }
             return true;
         }
+
+        // get inputs
         auto encoding = db.inputs.encoding();
         int width = static_cast<int>(db.inputs.width()); // Eigen types define scalar as int
         int height = static_cast<int>(db.inputs.height());
@@ -37,17 +39,17 @@ public:
         const double current_time = db.inputs.timeStamp();
         if (encoding == db.tokens.Type_RGB8)
         {
-            success = publishColorImageMessage<nvidia::gxf::VideoFormat::GXF_VIDEO_FORMAT_RGB>(
+            success = publishColorImage<nvidia::gxf::VideoFormat::GXF_VIDEO_FORMAT_RGB>(
                 db, height, width, dataAsCPU, current_time);
         }
         else if (encoding == db.tokens.Type_U8)
         {
-            success = publishColorImageMessage<nvidia::gxf::VideoFormat::GXF_VIDEO_FORMAT_GRAY>(
+            success = publishColorImage<nvidia::gxf::VideoFormat::GXF_VIDEO_FORMAT_GRAY>(
                 db, height, width, dataAsCPU, current_time);
         }
         else if (encoding == db.tokens.Type_U16)
         {
-            publishColorImageMessage<nvidia::gxf::VideoFormat::GXF_VIDEO_FORMAT_GRAY16>(
+            publishColorImage<nvidia::gxf::VideoFormat::GXF_VIDEO_FORMAT_GRAY16>(
                 db, height, width, dataAsCPU, current_time);
         }
         else if (encoding == db.tokens.Type_F32)
@@ -65,10 +67,11 @@ public:
 
 private:
     template <nvidia::gxf::VideoFormat Format>
-    static bool publishColorImageMessage(
+    static bool publishColorImage(
         OgnGXFPublishImageDatabase& db, int height, int width, const uint8_t* dataAsCPU, double time_seconds)
     {
         auto& state = db.internalState<OgnGXFPublishImage>();
+        // Create Camera message
         auto maybe_message = nvidia::isaac::CreateCameraMessage<Format>(
             state.getGxfContext(), width, height, nvidia::gxf::SurfaceLayout::GXF_SURFACE_LAYOUT_PITCH_LINEAR,
             nvidia::gxf::MemoryStorageType::kHost, state.mAllocator);
@@ -78,17 +81,94 @@ private:
             return false;
         }
         auto message = std::move(maybe_message.value());
+
+        // Populate timestamp
         message.timestamp->acqtime = static_cast<int64_t>(time_seconds * 1e9);
         message.timestamp->pubtime = message.timestamp->acqtime;
-        // GXF VideoBuffer is strided for some resolutions so we need to be careful here.
+
+        // Populate VideoBuffer
+        // Note: VideoBuffer is strided for some resolutions so we need to be careful here.
         const size_t dst_stride = message.frame->video_frame_info().color_planes[0].stride;
         const size_t src_stride = width * message.frame->video_frame_info().color_planes[0].bytes_per_pixel;
         cudaMemcpy2D(
             message.frame->pointer(), dst_stride, &dataAsCPU[0], src_stride, src_stride, height, cudaMemcpyHostToHost);
-        state.setIntrinsicsCamera(
-            message.intrinsics, width, height, db.inputs.focalLength(), db.inputs.horizontalAperture(),
-            db.inputs.verticalAperture(), db.tokenToString(db.inputs.projectionType()), db.inputs.cameraFisheyeParams(),
-            db.tokenToString(db.inputs.physicalDistortionModel()), db.inputs.physicalDistortionCoefficients());
+
+        // Populate intrinsics
+        const auto& focalLength = db.inputs.focalLength();
+        const auto& horizontalAperture = db.inputs.horizontalAperture();
+        const auto& verticalAperture = db.inputs.verticalAperture();
+        const auto& cameraFisheyeParams = db.inputs.cameraFisheyeParams();
+        const std::string projectionType = db.tokenToString(db.inputs.projectionType());
+        const std::string physicalDistortionModel = db.tokenToString(db.inputs.physicalDistortionModel());
+
+        if (physicalDistortionModel.find("rationalPolynomial") != std::string::npos)
+        {
+            auto fthetaWidth = cameraFisheyeParams[0];
+            auto fthetaHeight = cameraFisheyeParams[1];
+            auto fthetaCx = cameraFisheyeParams[2];
+            auto fthetaCy = cameraFisheyeParams[3];
+
+            message.intrinsics->dimensions.x = fthetaWidth;
+            message.intrinsics->dimensions.y = fthetaHeight;
+            message.intrinsics->focal_length.x = fthetaWidth * focalLength / horizontalAperture;
+            message.intrinsics->focal_length.y = fthetaHeight * focalLength / verticalAperture;
+            message.intrinsics->principal_point.x = fthetaCx;
+            message.intrinsics->principal_point.y = fthetaCy;
+            message.intrinsics->skew_value = 0.0;
+            message.intrinsics->distortion_type = nvidia::gxf::DistortionType::Polynomial;
+
+            const auto& physicalDistortionCoefficients = db.inputs.physicalDistortionCoefficients();
+            for (int i = 0; i < message.intrinsics->kMaxDistortionCoefficients; i++)
+            {
+                message.intrinsics->distortion_coefficients[i] = physicalDistortionCoefficients[i];
+            }
+        }
+        else if (projectionType.find("pinhole") != std::string::npos)
+        {
+            message.intrinsics->dimensions.x = width;
+            message.intrinsics->dimensions.y = height;
+            message.intrinsics->focal_length.x = width * focalLength / horizontalAperture;
+            message.intrinsics->focal_length.y = height * focalLength / verticalAperture;
+            message.intrinsics->principal_point.x = width * 0.5;
+            message.intrinsics->principal_point.y = height * 0.5;
+            message.intrinsics->skew_value = 0.0;
+            message.intrinsics->distortion_type = nvidia::gxf::DistortionType::Perspective;
+            for (int i = 0; i < message.intrinsics->kMaxDistortionCoefficients; i++)
+            {
+                message.intrinsics->distortion_coefficients[i] = 0.0;
+            }
+        }
+        else if (projectionType.find("fisheye") != std::string::npos)
+        {
+            auto fthetaWidth = cameraFisheyeParams[0];
+            auto fthetaHeight = cameraFisheyeParams[1];
+            auto fthetaCx = cameraFisheyeParams[2];
+            auto fthetaCy = cameraFisheyeParams[3];
+
+            message.intrinsics->dimensions.x = fthetaWidth;
+            message.intrinsics->dimensions.y = fthetaHeight;
+            message.intrinsics->focal_length.x = fthetaWidth * focalLength / horizontalAperture;
+            message.intrinsics->focal_length.y = fthetaHeight * focalLength / verticalAperture;
+            message.intrinsics->principal_point.x = fthetaCx;
+            message.intrinsics->principal_point.y = fthetaCy;
+            message.intrinsics->skew_value = 0.0;
+            message.intrinsics->distortion_type = nvidia::gxf::DistortionType::FisheyeEquidistant;
+            for (int i = 0; i < 5; i++)
+            {
+                message.intrinsics->distortion_coefficients[i] = cameraFisheyeParams[i + 5];
+            }
+            for (int i = 5; i < message.intrinsics->kMaxDistortionCoefficients; i++)
+            {
+                message.intrinsics->distortion_coefficients[i] = 0.0;
+            }
+        }
+        else
+        {
+            db.logError("Unexpected projection type: %s", projectionType);
+            return false;
+        }
+
+        // Populate extrinsics
         for (int i = 0; i < 3; ++i)
         {
             message.extrinsics->translation[i] = db.inputs.stereoOffset()[i];
@@ -97,6 +177,8 @@ private:
                 message.extrinsics->rotation[3 * i + j] = i == j ? 1 : 0;
             }
         }
+
+        // Publish message
         const gxf_result_t result = state.publish(db.inputs.outputEntity(), db.inputs.outputComponent(), message.entity);
         return (result == GXF_SUCCESS);
     }
@@ -131,19 +213,14 @@ private:
         return (result == GXF_SUCCESS);
     }
 
-    double getCurrentTime()
-    {
-        return mClock->time();
-    }
-
     void setMetadata(const double time_seconds, const nvidia::gxf::Handle<nvidia::gxf::Timestamp>& msgTimestamp)
     {
         msgTimestamp->pubtime = static_cast<int64_t>(time_seconds * 1e9);
         msgTimestamp->acqtime = static_cast<int64_t>(time_seconds * 1e9);
     }
 
-    void setIntrinsicsCameraImage(const nvidia::gxf::Handle<::isaac::geometry::PinholeD>& intrinsics,
-                                  const nvidia::gxf::Handle<::isaac::geometry::CameraDistortionInfo>& distIntrinsics,
+    void setIntrinsicsCameraImage(const nvidia::gxf::Handle<::nvidia::isaac::geometry::PinholeD>& intrinsics,
+                                  const nvidia::gxf::Handle<::nvidia::isaac::geometry::CameraDistortionInfo>& distIntrinsics,
                                   const int width,
                                   const int height,
                                   float focalLength,
@@ -164,94 +241,30 @@ private:
             physicalDistortionCoefficients.size() == 8)
         {
 
-            distIntrinsics->model = ::isaac::geometry::DistortionModel::kPolynomial;
+            distIntrinsics->model = ::nvidia::isaac::geometry::DistortionModel::kPolynomial;
             auto data = physicalDistortionCoefficients.data();
-            const std::array<double, ::isaac::geometry::CameraDistortionInfo::kMaxNumCoefficients> distortionCoefficients{
-                data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7]
-            };
-            distIntrinsics->distortion_coefficients = distortionCoefficients;
+            const std::array<double, ::nvidia::isaac::geometry::CameraDistortionInfo::kMaxNumCoefficients>
+                distortionCoefficients{ data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7] };
         }
         else
         {
             if (cameraFisheyeParams.size() == 19)
             {
                 auto data = cameraFisheyeParams.data();
-                const std::array<double, ::isaac::geometry::CameraDistortionInfo::kMaxNumCoefficients> distortionCoefficients{
-                    data[5], data[6], data[7], data[8], data[9]
-                };
+                const std::array<double, ::nvidia::isaac::geometry::CameraDistortionInfo::kMaxNumCoefficients>
+                    distortionCoefficients{ data[5], data[6], data[7], data[8], data[9] };
                 distIntrinsics->distortion_coefficients = distortionCoefficients;
             }
             else
             {
-                const std::array<double, ::isaac::geometry::CameraDistortionInfo::kMaxNumCoefficients> distortionCoefficients{
-                    0, 0, 0, 0, 0
-                };
+                const std::array<double, ::nvidia::isaac::geometry::CameraDistortionInfo::kMaxNumCoefficients>
+                    distortionCoefficients{ 0, 0, 0, 0, 0 };
                 distIntrinsics->distortion_coefficients = distortionCoefficients;
             }
 
             distIntrinsics->model = (projectionType.find("fisheye") != std::string::npos ?
-                                         ::isaac::geometry::DistortionModel::kFisheye :
-                                         ::isaac::geometry::DistortionModel::kPerspective);
-        }
-    }
-
-    void setIntrinsicsCamera(const nvidia::gxf::Handle<nvidia::gxf::CameraModel>& intrinsics,
-                             const int width,
-                             const int height,
-                             float focalLength,
-                             float horizontalAperture,
-                             float verticalAperture,
-                             const std::string projectionType,
-                             const omni::graph::core::ogn::const_array<float>& cameraFisheyeParams,
-                             const std::string physicalDistortionModel,
-                             const omni::graph::core::ogn::const_array<float>& physicalDistortionCoefficients)
-    {
-        auto fthetaWidth = cameraFisheyeParams[0];
-        auto fthetaHeight = cameraFisheyeParams[1];
-        auto fthetaCx = cameraFisheyeParams[2];
-        auto fthetaCy = cameraFisheyeParams[3];
-
-        intrinsics->dimensions.x = fthetaWidth;
-        intrinsics->dimensions.y = fthetaHeight;
-        intrinsics->focal_length.x = fthetaWidth * focalLength / horizontalAperture;
-        intrinsics->focal_length.y = fthetaHeight * focalLength / verticalAperture;
-        intrinsics->principal_point.x = fthetaCx;
-        intrinsics->principal_point.y = fthetaCy;
-        intrinsics->skew_value = 0.0;
-
-        if (physicalDistortionModel.find("rationalPolynomial") != std::string::npos &&
-            physicalDistortionCoefficients.size() == intrinsics->kMaxDistortionCoefficients)
-        {
-            intrinsics->distortion_type = nvidia::gxf::DistortionType::Polynomial;
-            for (int i = 0; i < intrinsics->kMaxDistortionCoefficients; i++)
-            {
-                intrinsics->distortion_coefficients[i] = physicalDistortionCoefficients[i];
-            }
-        }
-        else
-        {
-            if (cameraFisheyeParams.size() == 19)
-            {
-                for (int i = 0; i < 5; i++)
-                {
-                    intrinsics->distortion_coefficients[i] = cameraFisheyeParams[i + 5];
-                }
-                for (int i = 5; i < intrinsics->kMaxDistortionCoefficients; i++)
-                {
-                    intrinsics->distortion_coefficients[i] = 0.0;
-                }
-            }
-            else
-            {
-                for (int i = 0; i < intrinsics->kMaxDistortionCoefficients; i++)
-                {
-                    intrinsics->distortion_coefficients[i] = 0.0;
-                }
-            }
-
-            intrinsics->distortion_type =
-                (projectionType.find("fisheye") != std::string::npos ? nvidia::gxf::DistortionType::FisheyeEquidistant :
-                                                                       nvidia::gxf::DistortionType::Perspective);
+                                         ::nvidia::isaac::geometry::DistortionModel::kFisheye :
+                                         ::nvidia::isaac::geometry::DistortionModel::kPerspective);
         }
     }
 };